requirements.txt
librouteros~=3.2.1
python-dotenv~=1.0.1
matplotlib~=3.9.2